+++ boot/scripts/02_setup.sh
@@ -87,13 +87,25 @@ fi
 # patch resolvconf.conf for local DNS and domain name
 try_patch "/etc/resolvconf.conf" "local name servers"
 
-# sysctl customisations 
+# sysctl customisations (default disables IPv6)
 # Step 1: check for sysctl.conf and apply it if found. That edits
 #         /etc/sysctl.conf directly. This is the legacy approach.
 # Step 2: check for sysctl.d. If it is found, copy any files found
 #         inside it (no overwrite). This is the modern approach.
 try_patch "/etc/sysctl.conf" "patching sysctl.conf"
 try_merge "/etc/sysctl.d" "customising sysctl.d"
+
+# grub customisations for hosts booting that way. Raspberry Pis don't
+# use grub so this is aimed at native/ProxMox Debian/Ubuntu. The
+# default provided with PiBuilder disables IPv6
+if try_merge "/etc/default/grub.d" "customising grub.d" ; then
+   if [ -x "/usr/sbin/update-grub" ] ; then
+      echo "Updating GRUB"
+      sudo update-grub
+   else
+      echo "Warning: PiBuilder patched /etc/default/grub.d but /usr/sbin/update-grub not present"
+   fi
+fi
 
 # patch journald.conf to reduce endless docker-runtime mount messages
 try_patch "/etc/systemd/journald.conf" "less verbose journalctl"